Backup and restore .git directory during middleman builds
Middleman builds are now fully "clean" by default, but this means that
the .git directory that we first initialize in the build directory gets
automatically removed by `middleman build` after we first sync with the
gh-pages branch.

This change moves the .git directory to a temp directory before the
middleman build and then moves the git directory back immediatey
afterwards allowing us to properly track changes, commit, and push to
the gh-pages branch.